Trim a character vector

Almost identical to glue::trim() save a slight difference in error handling for non-character input. This function trims a character vector according to the trimming rules used by glue. These follow similar rules to Python Docstrings, with the following features:

Leading and trailing whitespace from the first and last lines is removed.

A uniform amount of indentation is stripped from the second line on, equal to the minimum indentation of all non-blank lines after the first.

Lines can be continued across newlines by using \\.

Usage

trim(x)

Arguments

x

[character].

Value

A character vector.

See also

Examples

cat(trim("
    A formatted string
    Can have multiple lines
      with additional indentation preserved
    "))
#> A formatted string
#> Can have multiple lines
#>   with additional indentation preserved

cat(trim("
  \ntrailing or leading newlines can be added explicitly\n
  "))
#>   
#> trailing or leading newlines can be added explicitly

cat(trim("
    A formatted string \\
    can also be on a \\
    single line
    "))
#> A formatted string can also be on a single line